LPS-72396 Move to match com.liferay.expando.kernel.util
